# backend-app/binary/http/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Comprehensive HTTP API for URL shortener, graph algorithms, friend management, messaging, and user authentication",
        "title": "Portfolio API",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/msyamsula/portofolio"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:5000",
    "basePath": "/",
    "paths": {
        "/friend/add": {
            "post": {
                "description": "Adds a friendship relationship between two users",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "friend"
                ],
                "summary": "Add friend",
                "parameters": [
                    {
                        "description": "Add friend request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_msyamsula_portofolio_backend-app_domain_friend_dto.AddFriendRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_msyamsula_portofolio_backend-app_domain_friend_dto.AddFriendResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/friend/get": {
            "get": {
                "description": "Retrieves all friends for a given user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "friend"
                ],
                "summary": "Get friends",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_msyamsula_portofolio_backend-app_domain_friend_dto.GetFriendsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/graph/solve/{algo}": {
            "post": {
                "description": "Executes the specified graph algorithm on the provided graph",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "graph"
                ],
                "summary": "Solve graph algorithm",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Algorithm name (dfs, bfs, cycle, dag, scc, ap, ep)",
                        "name": "algo",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Is graph directed",
                        "name": "isDirected",
                        "in": "query"
                    },
                    {
                        "description": "Graph notation",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_msyamsula_portofolio_backend-app_domain_graph_dto.SolveRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_msyamsula_portofolio_backend-app_domain_graph_dto.SolveResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_msyamsula_portofolio_backend-app_domain_friend_dto.AddFriendRequest": {
            "type": "object",
            "properties": {
                "big_id": {
                    "type": "integer"
                },
                "small_id": {
                    "type": "integer"
                }
            }
        },
        "github_com_msyamsula_portofolio_backend-app_domain_friend_dto.AddFriendResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_msyamsula_portofolio_backend-app_domain_friend_dto.GetFriendsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_msyamsula_portofolio_backend-app_domain_friend_dto.User"
                    }
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_msyamsula_portofolio_backend-app_domain_friend_dto.User": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "online": {
                    "type": "boolean"
                },
                "unread": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "github_com_msyamsula_portofolio_backend-app_domain_graph_dto.Edge": {
            "type": "object",
            "properties": {
                "from": {
                    "type": "string"
                },
                "to": {
                    "type": "string"
                },
                "weight": {
                    "type": "string"
                }
            }
        },
        "github_com_msyamsula_portofolio_backend-app_domain_graph_dto.GraphNotation": {
            "type": "object",
            "properties": {
                "edges": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_msyamsula_portofolio_backend-app_domain_graph_dto.Edge"
                    }
                },
                "nodes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "github_com_msyamsula_portofolio_backend-app_domain_graph_dto.SolveRequest": {
            "type": "object",
            "properties": {
                "graph": {
                    "$ref": "#/definitions/github_com_msyamsula_portofolio_backend-app_domain_graph_dto.GraphNotation"
                }
            }
        },
        "github_com_msyamsula_portofolio_backend-app_domain_graph_dto.SolveResponse": {
            "type": "object",
            "properties": {
                "acyclic": {
                    "type": "boolean"
                },
                "ap": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "bridge": {
                    "type": "array",
                    "items": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "cycles": {
                    "type": "array",
                    "items": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "log": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "path": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "scc": {
                    "type": "array",
                    "items": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer token authentication",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}